requirements.txt
langchain==0.1.14
langchain_community==0.0.31
langchain_experimental==0.0.56
langchain_google_genai==1.0.1
langchain_text_splitters==0.0.1
pymupdf==1.24.1
python-dotenv==1.0.1
streamlit==1.33.0
tabulate==0.9.0
